Minor Bug Fixes
Fixed some additional warnings the VS compiler gave out and edit the
CMakelists to better differentaite between WIN32 and OSX/Linux compiler
options

+++ engine/src/EngineApp/EngineApp.cpp
@@ -37,7 +37,7 @@ void EngineApp::run(Scene * scene, I_Renderer * renderer)
 	//Set the global active scene to this one.
 	SetActiveScene(scene);
 	initializeComponents(scene);
-	float newT, t = glfwGetTime();
+	float newT, t = (float) glfwGetTime();
 	float dt = 0;
 	renderer->setup(scene);
 	while(!glfwWindowShouldClose(window))
@@ -48,7 +48,7 @@ void EngineApp::run(Scene * scene, I_Renderer * renderer)
 		scene->runCollisionUpdate();
 		scene->runUpdate(dt);
 		renderer->render(scene);
-		newT = glfwGetTime();
+		newT = (float) glfwGetTime();
 		glfwSwapBuffers(window);
 		dt =  newT - t;
 		t = newT;
