+++ engine/src/Scene/Scene.cpp
@@ -64,9 +64,9 @@ void Scene::runCollisionUpdate()
 {
 	if(_boxCollisionComponents.size() > 0)
 	{
-		for(int i = 0; i < _boxCollisionComponents.size() - 1; i++)
+		for(size_t i = 0; i < _boxCollisionComponents.size() - 1; i++)
 		{
-			for(int j = i + 1; j < _boxCollisionComponents.size(); j++)
+			for(size_t j = i + 1; j < _boxCollisionComponents.size(); j++)
 			{
 				//Try collision between i and j
 				if(_boxCollisionComponents[i]->intersects(
